+++ extension.ts
@@ -9,6 +9,8 @@ import { exportAll, exportExplorerItem } from './commands/export';
 import { ObjectScriptClassSymbolProvider } from './providers/ObjectScriptClassSymbolProvider';
 import { ObjectScriptRoutineSymbolProvider } from './providers/ObjectScriptRoutineSymbolProvider';
 import { DocumentContentProvider } from './providers/DocumentContentProvider';
+import { ObjectScriptClassFoldingRangeProvider } from './providers/ObjectScriptClassFoldingRangeProvider';
+import { ObjectScriptFoldingRangeProvider } from './providers/ObjectScriptFoldingRangeProvider';
 
 import { ObjectScriptExplorerProvider } from './explorer/explorer';
 import { outputChannel } from './utils';
@@ -82,7 +84,12 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
     vscode.languages.registerDocumentSymbolProvider(
       { language: 'objectscript' },
       new ObjectScriptRoutineSymbolProvider()
-    )
+    ),
+    vscode.languages.registerFoldingRangeProvider(
+      { language: 'objectscript-class' },
+      new ObjectScriptClassFoldingRangeProvider()
+    ),
+    vscode.languages.registerFoldingRangeProvider({ language: 'objectscript' }, new ObjectScriptFoldingRangeProvider())
   );
 }